Remove the species size factor for SH
For kite shields, large species got -0.5 SH and kobolds got +0.5 SH.
For bucklers, kobolds got +1 SH and spriggans got +2 SH. This almost
didn't matter and was very silly. The intent, I believe, was to
simulate that shields covered a smaller or larger part of the wearer's
body, but we don't really need to have a tiny bonus modifier for that.

(It was also, I believe, totally undocumented.)