Rework storm shield part.3
The noise from the clouds in the early game or abyss forced the player to engage in an overly harsh battle. I used the wereblood method to reduce the 'harshness' of storm shields, (but still make a lot of noise.) The storm subsides when you move the floor. But if you kill an enemy, the storm's figure does not decrease until you move the floor.

Add rough dust cloud:
The dust cloud does not cause any damage and wastes the opportunity for other elemental clouds to be created. I want players to endure a long battle instead of dancing on the stairs. So I'm going to add a new cloud that does physical damage.

ac bonus: This option added instead of sh bonus.