Add labels to unrand artefact effects
To make special effects 'pop' more, make it more obvious when an
unrand does something unusual, and correlate inscriptions with their
effects, add inscriptions for various unrand effects that didn't have
any and preface unrand effect descriptions with a corresponding
inscription. This is incomplete and a little ugly but I'm tired of
staring at unrands for now, sorry.