Remove Morg
This has always been an extremely unexciting unrand, both from a lore
and a gameplay perspective. Moving it to the early game didn't change
that. We can add it back if we ever find a good gimmick for it, but
we've been trying for years without any success, so I'm skeptical.