Give ?butterflies knockback (acrobat)
To make its effect more comparable in strength to ?summoning, which
it now alternates with, apply a powerful knockback to nearby
creatures. Unlike other forms of knockback, this does no damage,
mostly so that we don't have to worry about prompting the player.
("Really knock back the goblin in sanctuary..?")