feat: toggle equip vs unequip modes in UseItemMenu
This commit adds a key (bound right now to tab) that, in the equip
menus, toggles to the corresponding unequip menu (or general unequip if
there isn't one) and vice versa. This is a bit overengineered, but the
end goal of this is that a new player only needs to remember one key
to get to all equip menus from the main view (`e`).

Note: the option `equip_unequip` also can accomplish this goal (in
combination with the new equip menus). However, it makes it a bit too
easy to accidentally unequip stuff, and also (wheals) new players mostly
don't want to unequip stuff in modern crawl, so there's not a huge need
to make it so immediately accessible.

Other notes:
* this commit attempts to do smarter hover saving on mode changes; it
  doesn't just preserve position, but item link if possible. This
  applies to all menu mode changes.
* keyhelp code here is kind of involved...
* commit should have no effect outside of equip/unequip menus aside from
  more consistent keyhelps.