decouple gold aura duration and corpse size
Killing bigger monsters gave you longer gold aura duration, which is
mechanically weird (also not great that you could pseudo-count the duration
by looking at the size of a gold stack that dropped!) Just roll independently
instead, choosing numbers roughly equal to the case for human size. Gold amount
being tied to corpse size is also weird in a post-chunk universe, but the
gameplay implications there are mostly fine and it would be a significant
balance change to decouple it - worth looking at if gozag needs another nerf,
but I've left it alone for now.