Warn before launching inappropriate martial attacks
Previously, the game asked for confirmation before attacking with
(e.g.) a weapon inscribed with !a. It also did this before rampaging
into a monster, but not for Wu Jian's "martial attacks".

It now carries out the same checks before Wu Jian's lunge and whirlwind
attacks, and does it before a wall jump if this attacks a monster.

Closes #2962.