Add Sling of Outrageous Fortune
Adds a new unrand sling.

From a gameplay standpoint, the design goals of this item are to interact interestingly with slay, and to achieve the fun of big !!! while having a very different feel from Dark Maul.
From a flavor standpoint, the entire point of this item is the Hamlet reference.