feat: Support up to 100 waypoints
For establishing a user coordinate system on levels on the clua side, we
need to be able to create at least one waypoint per permanent level.
Crawl has almost 90 such levels, so 100 waypoints will cover these plus
the extra waypoint needed for temporary levels such as portals, Pan, and
the Abyss.

This does add 90 waypoint slots to the save file and to travel memory,
but at a cost of 9 bytes per waypoint (one short for a level ID and two
ints for a coordinate), this doesn't matter. This commit just extends
the allowed number of waypoints without allowing the in-game UI to add
waypoints past the 10th. The additional waypoints can be created via
clua's `travel.set_waypoint()` function and additionally deleted in the
UI along with all other waypoints by using the "*" option.