fix: add some arena item checks
These functions all assume a reasonable value for `you`, which can't be
assumed in the arena. These three checks have some overlap for the
reported crash, but it seems good to cover a bunch of bases here since I
think multiple code paths code apply, and some of the checks in this
code that don't crash may still have undefined behavior.

Resolves #2955