fix: fully disallow python 2, tweak deprecation handling
Python 2.7 was deprecated with the 0.29.0 release in Aug, and now
crashes on start because of the use of a py33 module (py33 was released
in 2012). It would probably be possible, in principle, to get py27
working again, but nearly all servers have updated, and I'm pretty tired
of extreme backcompat hacks. This commit prevents the server from
starting either with an old version of python, or with an old version of
Tornado. In the future some backcompat commit can probably also be
cleaned up, since I'm not sure this affects more than ~1 server.