fix: better handle a webtiles chroot error case
A common webserver setup issue (on a new setup, or an upgrade from an
old version of tornado) is that modern webtiles under a chroot requires
a copy of the python library in the chroot. This is because any `import`
after the chroot call is still looking for the python library at the
same path as before, but paths have changed; older tornado (and a lot of
older python code in general) simply didn't use locally scoped import.
However, it is good python practice in a lot of cases, and are now
used both in modern tornado and various parts of the webtiles server.

This usually seems to fail first in tornado code (though I think that's
not the only possibility) -- so this commit adds an explicit check for
this and attempts to provide a more informative error message along with
the exception. By its nature this is a bit hard to directly test
locally...