New unrand: Storm Queen's Shield
The Disc of Storms is back, in completely unrelated form!

Alternatively: did you ever want to be a shock serpent?

The Storm Queen's Shield is a +5 kite shield of flight and rElec that
has a chance to retaliate on block, lashing out with 3d7 of
half-AC-checking electric damage. This effect triggers even up to
range 3 (as with shock serpents). It's a bit like reflection, but it
works on melee blocks (unlike reflection).

Not sure this ended up being extremely exciting, but I did implement
it, so let's see!