Halve the chance of Irradiate malmutation
To focus a bit more on the exciting parts of the spell (damaging
enemies & contaminating the player), malmutate victims only on a
coinflip chance. We could play with these numbers further as
appropriate - for example, by making the chance depend on power
and/or victim HD, or just by adjusting the flat rate.