New monster: skyshark
Slicing through the Dungeon's air - oh, what graceful beasts! What
a terrifying sight...!

Skysharks are a speed-15 flying monster that becomes more dangerous
once it tastes blood, first getting Might (+50% damage) and then
going Berserk. Perhaps wise to kill them before that happens?

They're currently intended as a high-end threat for Lair and a low-end
Shoals monster, but they might migrate to the Desert once that lands.
So to speak.

Canonically, this species is related to, but distinct from, the skysharks
that the Slick Slippers are made from.