refactor: rewrite maybe_bool
This commit rewrites maybe_bool to be, argubly, better. See below for
motivation, but here are the practical notes up front:
* `maybe_bool` becomes a class rather than an enum.
* After this commit, nearly anywhere you would write MB_FALSE or
  MB_TRUE you would simply write `false` or `true`.
* Where you would write MB_MAYBE you now should write
  `maybe_bool::maybe`.
* Comparison to `true` or `false` is safe and will do exact
  comparison. (The latter may appear surprising, but it is because
  comparison will cause a bool to be cast to maybe_bool, rather than a
  maybe_bool to be cast down to bool.)
* Semantics change: `maybe_bool::maybe` does not convert to true by
  default under any circumstance.
* In cases where you would have previously written `mb == MB_TRUE` you
  can just rely on a bool cast (`maybe` does not convert to `true`). In
  cases where you would write `mb == MB_FALSE` you can write `!mb`.
* frombool and tobool are replaced respectively by constructors and
  the class method maybe_bool::to_bool (as well as an explicit bool
  cast).
* While bool explicit operators have weird rules and can be used
  implicitly in some cases, there are still times when you'll need
  to do the cast. In particular, if a function returns a bool, to
  return a maybe_bool you'll need to explicitly cast.
* Mixing bools and maybe_bools in logical expressions generally requires
  an explicit cast. (If you write a condition that just uses !, that
  doesn't need one.)

Old `maybe_bool` had the advantage of simplicity: it was just an enum
and as long as these enum values are used consistently, it is pretty
simple to use and as efficient as any integer-based type, with
effectively no implementation code.

However, there are a few complaints about it. First, the naming scheme
was quite ad hoc across the enum values and various functions that used
them.  Second, and more importantly in my mind, the enum-based
implementation led in a very non-clean way to implicit int and bool
casts, and a very ad hoc semantics for this type, especially in its
interaction with boolean expressions. Taken as a three-valued logic, it
is very weird. The law of the excluded middle *is* valid (in the sense
that `p || !p` is always true), but double negation elimination is *not*
valid (i.e. `!!p != p`); the weaker double negation inference `!!!p ==
!p` does hold conceivably classing it with some paraconsistent logics,
though I haven't found an exact analogue (see [1], though this logic
doesn't much resemble L8 there, which has the same double negation
pattern). At a practical level this ad hoc nature made mixing with bools
in both directions weird and somewhat error prone; aside from the above
concerns (which essentially result from MB_MAYBE implicitly casting to
`true`), `true` would cast to the enum value MB_MAYBE, leading to clunky
conversion code in a lot of places, needing to always use MB
constants, and error-prone comparison with regular bools.

This commit tries to systematize things more cleanly, albeit with a much
more complicated implementation. The underlying data type does still
amount to a fancy enum (now an enum class), but it's wrapped with a more
clearly defined class, largely inspired by boost's logic::tribool.  (I
considered whether importing one of the various "`optional` but for
c++11" packages out there might be a solution, but ultimately rejected
this, because these classes tend to have an even weirder, for a
bool-based type, set of implicit conversions to bool). A `bool`
constructor is provided making it generally safe to use `true` and
`false` as e.g. return values, and an explicit cast to bool is provided
that has a better default than MB_MAYBE->true. (In some cases, this may
be the desired semantics, of course, but this is still doable via
`to_bool` or direct checks against false.) The operator semantics is
based on the standard "Strong Kleene" three-valued logic operators,
which provide a (imo) very sensible set of defaults for a three-valued
logic. Note that because of the existence of the bool operator, this
won't let you mix `maybe_bool`s in logical expressions together with
bools without some explicit casting one way or the other.  The
implementation is nearly (but not quite) purely header-based.

As part of this commit, I have implemented everything as a
back-compatibility layer and purely just tried to clarify the semantics
of various instances of maybe_bool in cases where it did get an implicit
conversion. The bulk of this is around `can_wear_item` which does tend
to treat its maybe return as truth-y.  A future commit will convert
MB_TRUE etc into their `maybe_bool::t` etc equivalents.

[1] Kamide 2013, A Hierarchy of Weak Double Negations, Studia Logica 101.6

fix