feat: Better flight reasoning for lua functions
For the clua functions travel.feature_traversable() and
view.is_safe_square, add an assume_flight boolean argument that has
these assume the player has permanent flight. It's important to be able
to reason about traversability and safety when considering using
temporary flight or swapping to a permanent flight item. It wouldn't be
possible to do use these functions to reason about flight without an
argument like this, since there are multiple reasons a square can fail
to be traversable or safe.

These functions could instead return a table indicating all the reasons
why a square fails to be safe or traversable, but that would break
current scripts, so we'll do this for now. The comments for these
functions have been updated to more accurately describe what they test.

This is a minimal feature that I've tested and will be pushing to
stable. I'm making it now for the purpose of running qw during
tournament. It adds a new argument to `feat_is_traversable_now()` on the
C++ side, but with a default value that gives identical behaviour to the
old function.

(cherry picked from commit 5d4078a31ebbdd78447a0ff88a57d9623ba68f84)