fix: disallow unequipping cursed weapons
It was possible to unequip and destroy a cursed weapon by simply
switching to unarmed via `e-`. Similarly, this could be done in the
unified unequip menu by pressing `c` and the weapon's letter.

Compared to Shatter the Chains, shattering an item by unequipping it
took 3 auts instead of 10, and it could be done even under silence.
Also, the player could unequip the obsidian axe while mesmerized by
an enemy.