fix: don't allow cursing/uncursing wrong items
Although `prompt_invent_item()` supports filtering items, it doesn't
check if the selected item is appropriate. Curse Item and Shatter
the Chains abilities don't have such checks either, so it was possible
to use them on wrong items.

The player could apply curses to non-equipped inventory items using
Curse Item. Similarly, it was possible to press `*` and select a
totally inappropriate item, e.g. a potion, when destroying a curse,
which crashed the game.

This commit fixes these bugs by ensuring that the item returned by
`prompt_invent_item()` is indeed cursable and equipped when using Curse
Item, or that it's a worn cursed item when using Shatter the Chains.

Note that unlike `item_is_cursable(item)`, `item_is_selected(item,
OSEL_CURSABLE)` also checks that the item is equipped.