Make condenser vane good god approved
Good gods currently get upset only when condenser vane rolls Negative Energy clouds.
The chance of rolling Negative Energy clouds is somewhere in the vicinity of 1/1000 at 0 evocations, rising dramatically as evo skill increases.
This creates an odd incentive to not skill evocations.

Squashing negative energy clouds (as opposed to rerolling) was chosen to be the minimum change that resolved the problem.