feat: A lua function to test known map boundaries
This adds the clua function `view.in_known_map_bounds(x, y)`, which
returns true when the given coordinates are in the bounding box of map
areas known to the player. This is necessary to call functions like
`travel.is_excluded()` without generating an error, since this and
similar functions make such a check. Callers of these functions could
just use `pcall()` to trap the error, but it's good to have parity
between information the UI provides and what clua can determine.

This is a minimal feature addition that I've tested and will be pushing
to stable. I'm making it now for the purpose of running qw during
tournament. It turns the formerly static `_in_map_bounds()` to
the global `in_known_map_bounds()`, but the behaviour is identical. I've
also documented the relevant crawl-side C++ functions.

(cherry picked from commit 282de0ab27e44dfef9c7de68589d803a3389b590)