Convert the staff of Dispater into an orb
Of the Hell/Pan lord unrands, the staff of Dispater is one where there's at
least a small chance that a character strong enough to acquire it might
still be able to make use of it, with evokable damnation being fairly
powerful. However, using it means swapping weapons repeatedly to activate
the ability. Convert it to an orb instead, and give it *Corrode and some
bonus AC for some minor added flavour (such that the AC compensates for when
corrosion kicks in, but with some hellish flavour and a downside along the
lines of regular orbs).

No tile yet!