fix: (quick fix) disable a bad android build line
This no doubt breaks the android build to remove, but it *really really
really* should not be attempting to rm -f -r this directory! (which in a
server build context, might be /var/lib/dgamelaunch or something --
luckily CDO has permissions set so this didn't work, hopefully true of
any other servers configured like this). I'm not sure what the best fix
for someone who is actually working on the android build process might
be.