refactor: convert the travel prompt to a menu
This is a somewhat involved refactor that should mostly have no impact
on the use of this menu in regular play. Previously, the travel prompt
simply printed a bunch of stuff in the message pane and prompted. In
most normal cases this is fine, but it was not robust to overflowing the
message pane, resulting in either a more, or lines that were simply not
invisible. (The case that reliably hits this is wizmode &~, which is
super annoying unless you have the branch hotkeys memorized, which I
really, realy don't.)

This commit implements a new Menu subclass that can format its menu
entries to be shown in a message pane prompt, and run a regular input
loop in the message pane. If the menu fits, it does this, otherwise, it
uses a menu popup. The travel prompt now uses this feature, which also
nicely separates formatting from menu logic. It's quite a bit more code,
though.  There's only one intended behavior change to this menu, which
is to support tab as a "last destination" hotkey -- enter won't work
when this menu is a popup. Otherwise, everything should behave pretty
much the same as before when this menu is in prompt mode.

The wizmode command is still a bit annoying because it doesn't fit every
branch on one screen still in 80x24 console, but I have a planned future
commit to address this.

(In the longer run, a planned use case for this sort of approach is to
add options to make some message pane input more mouse/touchpad
friendly.)