feat: New Unique Regal and New Monster Spell Charming Aura
Regal is intended as a late game monster that provides a different ally
counter than mass abjuration.  The front rank of allies have a chance to
get charmed which turns them against the back ranks.  Summoners can then
decide to recast hoping for it to be closer (the charmed summons still
exists as a hostile).  I'm hoping the dynamic makes the fight more
tactical for the summoner.  Without the entourage he ends up being
rather bland and only an interesting fight for dedicated summoners, but
with the Klown and cob it should even push the blaster caster away from
the otherwise no brainer servitor into a more dynamic fight.  The
inclusion of an item from the Octopus King set is to flavor the octopod
unique and this late in the game should not be overpowering.

New Unique: Regal

A unique Octopode prince carrying one unique Octopus King item and able
to turn most foes whose gaze falls directly on him to his side.  He
spawns with a small entourage of a Killer Klown and a Death Cob to
increase the threat level.

New Monster Spell: Charming Aura
Attempts to charm everything in direct line of fire that isn't blocked
by other monsters.  This is not a full screen effect like Olgreb's Toxic
Radiance; it only charms the front rank of monsters by firing tracers
against all monsters in LOS_NO_TRANS and discarding failures.