feat: a console multicolumn menu implementation
This commit generalizes the implementation of MF_USE_TWO_COLUMNS for
menus beyond just local tiles; two-column menus on console now work as
well. This is kind of a basic implementation in some ways, and e.g. it
will still use multiple columns if item names exceed the column width
(for local tiles, it will actually wrap such item names). But, it works
fairly well, and this is mainly an issue for artefacts whose name can be
viewed by looking at the item.

Subsidiary changes / notes:
* the relevant option is given a more general name (with the old one as
  an alias)
* the relevant option is explicitly applied only to item menus. The main
  case this affects is large search results in the help lookup menu,
  which in my testing looked pretty good as multi-column results.
* no effect on webtiles