fix: prevent a display bug in multicolumn multiline tiles menus
The `do_layout` call for UIMenu is very side-effect-y. It's also used in
the _get_preferred_size calculation for this class. But (for reasons
that I don't fully grasp) the ui code will call _get_preferred_size at
all sorts of strange times with quite speculative values. This may
itself be some sort of bug, I'm not really sure; but the side effects
were leading to rendering bugs. In particular, sometimes
_get_preferred_size would be called with the max window width, which
leads to rendering inventory items on a single line in a dense
multi-column layout, and the line positions for this rendering choice
would be cached. If `pack_buffers` gets called before a fresh do_layout
call with the correct width happens, the menu spacing changes because of
the cached values. For unclear reasons, this often happens on a
pagedown, and also in various scrolling circumstances.

This commit simply attempts to limit the side effects that matter during
_get_preferred_size calls, rather than looking too deeply into the `ui`
whys.