fix: tweak multicolumn scroll behavior
If the hover is in a position where the menu can scroll, but the hover
itself can't scroll (typically, it's in the right-hand column with a
one-item section directly below as the last section), a down/up arrow
would not do anything. This leads to the menu feeling kind of stuck --
it makes intuitive sense that the hover doesn't move, but the menu has
further items that just don't get revealed. This commit adds a scrolling
behavior for this case even with the hover not moving. In my testing it
feels fairly natural, much better than the old behavior, but it's a bit
ad hoc, so we'll see.