Adjust Harm display on the % overview screen
As with other properties that are rarer/artefact-only, display Harm only if
the player has any levels of it. Display it with two pips so that the fact
that Harm has exactly two levels is visible to the player, move it and
*Rage to the bottom (as the two properties in the second column that aren't
just binary), and shift the equipment column over by a character to make a
little more room.

There are some ideas for a larger rework of the % screen, but in the
meantime this should hopefully be an improvement.