Simplify the spectral brand
To make it easier to understand how the spectral brand works, rather
than using the stats of a strange 'spectral weapon' monster to determine
accuracy and damage, use the player's stats instead. (Then reduce pre-AC
damage to 70%, matching cleaving.)

This should make the brand scale more reasonably over the course of the
game (especially in the early game), and make it easier to make future
balance adjustments as needed.