Remove 'failure to reach'
This effect gave a 50% chance of failure when using a polearm to attack
over allies' heads. It was frustrating, undocumented, confused players,
and applied to an increasingly small fraction of allies. Why should
followers of Beogh, Yred or Jiyva get a 50% fail chance when Hep or Fed
worshippers didn't?

Reaching over allies is a strong effect, but it's my suspicion that it
won't be dominant even if this 50% fail chance is removed. Let's find out.