refactor: better separate option parsing, state, and management
Crawl options are extremely involved, with a massive set of options that
are set up and parsed in a wide variety of ways, with nearly all the
work being centralized in one heavyweight class (`game_options`) that is
primarily instantiated as a singleton. As this code has grown it has
become increasingly challenging to maintain it, especially when
considering significant user-facing improvements to configuration.

This commit attempts to impose (what I think is) a better structure on
this. I have split `game_options` into a base class `base_game_options`,
that is responsible for managing option "meta-state", and where generic
option-related code goes; option parsing is now run from this class. The
`game_options` class is now responsible for actual options fields, as
well as custom parsing. I haven't attempted (yet) to eliminate the giant
custom parsing conditional sequence, but it is now partitioned off from
the general parsing code. As part of this I have abstracted option
parsing state out of the parsing function itself, allowing parsing to be
split across multiple functions in a cleaner way.

There were probably multiple ways to achieve these refactoring goals,
but this one (a) keeps api compatibility for Options for external users,
and (b) makes it possible to write a sane copy constructor / assignment
operator for the meta-state (especially the vector of `GameOption *`s)
while using a defaulted copy constructor for the actual options fields.
Previously a defaulted copy constructor existed, but it wasn't safe to
use for any purpose except accessing the raw fields, because the
metastate wasn't handled correctly. The existing `merge` function
couldn't substitute for general purposes because it can only handle
GameOption style options.

As a refactoring commit, this leaves things in a somewhat intermediate
state: I have done what is necessary to get the top-down changes to
work, but there's a lot of messy code here still..