Remove some iron grates from evilmike_geh (#2980)
They seem to exist to allow players to dig out and fly across some areas of
the lava, but could result in permaflying players getting trapped by
teleporting into the lava area without a way to get through grates. The
areas with grates also place stokers, which are theoretically able to
summon through the grates but in practice were unlikely to do so
successfully - removing the grates and opening up those areas allows the
outer lava monsters a way in.

Closes #2980.