refactor: generalize option case handling
Before this commit, case sensitive GameOption-based options needed
custom handling, because a `lowercase` was done on the string used for
the constructor. This adds support for case sensitivity to a couple of
GameOption subclasses, and uses line parser state in the constructor so
that these classes can have access to the case sensitive version. As
part of this, I was able to fully remove the hardcoded list of case
sensitive options. Sadly, only one custom parsed option was able to be
generalized as a result of this.

Also, this fixes what seems to be a bug where morgue_dir is not treated
as case sensitive.