Tweak holy_pan to remove teleport closets (#2979)
Historically, Pan vaults got away with having teleport closets under the
assumption that anyone going there would have plenty of means of flight,
digging and/or teleportation, but it's also pretty simple to avoid in this
case. Move one of the of the exits to the opposite side of the river, tweak
some terrain to open up a pathway along one side of the river, and make a
couple of spots (Elyilon's altar island and the pearl dragon cave)
no_tele_into.

Closes #2979.