fix: show rDrown and rBlind in the monster lookup (?/m)
Previously, these resistances were displayed only when inspecting a
monster via `xv`. But, for example, `?/mjelly` didn't mention rBlind
and rDrown at all.

In-game monster descriptions are constructed using a `monster *`.
On the other hand, monster lookup produces descriptions using just
the type and base type of a monster. The latter constructor initialized
the basic resists, but it didn't check for rBlind nor rDrown.

There was no way to check drowning resistance using only the type of
a monster, so this commit adds a new utility function for that.