refactor: generalize some bitfield list options
These three options are generalized using a more invasive technique than
previous commits, to handle an underlying bitfield type. Instead of
trying to convert the list handling directly into a templated
ListGameOption or writing a new options class, I have stored the list as
a string, and triggered an update function each time the string is
changed, which updates the real value by overwriting the bitfield.  This
has the downside that if the actual option state is changed outside of
the option handling code, it does not trigger a change in the list
option state; but currently this does not matter.