refactor: convert map scale options to FixedpGameOption
This standardizes these two options using a new option class that can
handle decimal values directly, using `fixedp<>`.

...and, uh, this commit merges fixedp.h. This commit would at very long
last, resolve #810. Though, unlike the original idea of that PR, I no
longer think there's a need to wholesale convert scaled int code to
fixedp in order to merge!  It is still a really handy class to have
around for a lot of crawlcode, and it's kind of silly that it has sat
for so long without seeing any use.

(Insider notes: this version of fixedp.h has a few tweaks beyond the
2018 version. Primary ones are: integrating the test code with
crawlcode, and implementing numeric_limits min/max. Also added an
explicit copy constructor and dealt with various new warnings. We'll see
how it does on gcc...)