fix: prompt menu webtiles formatting (dilly)
Unfortunately there seems to be a mismatch in how webtiles is handling
multiline messages, where it should insert a space before every line
that doesn't have a prefix, but isn't. Looking at the webtiles code this
seems quite hard to fix, so I've retreated to individually printing the
prompt menu lines. (This is how the old `G` menu did this, which is why
it didn't have this bug before.)