fix: rewrite the `enemy_hp_colour` option
The code that uses this option is hardcoded to assume a particular size
vector with each position having a designated value. However, the option
implementation treated it as a list, with complete flexibility as to
size, allowing prepending, subtraction, etc., and no validation of the
vector size. I haven't gotten it to crash but suffice to say the
behavior of the consumer code is at best undefined if the option is not
set correctly. It looks to me like the list manipulation patch 11 years
ago didn't look closely enough at what this option did when implementing
prepending etc for this option, though the lack of size validation goes
back to the original implementation of this option in ff9fb87a6f8b8d.

This commit reimplements the option behavior to always use a fixed size
vector. If not enough options are supplied, the unsupplied values remain
default, and if too many options are supplied, it produces an error.
I've also added the possibility of writing "default" for any of the
slots as a convenience.