refactor: move some custom options to base_game_options
The `macros` and `bindkey` syntaxes aren't appropriate for converting to
a GameOptions implementation, and in addition, the former already has an
in-game editing UI (the latter probably should, but doesn't; if it does,
it'll be distinct from a regular options-editing UI).
So, move them to the base class
list of manually handled meta-options (which is where I'm keeping things
that have this property).

Includes are handled in the base class. So move `additional_macro_file`
to the base class as well. (I don't know if turning this one into a
ListGameOption would be viable, but it might be good to generalize the
syntax..)

Also, fix some reset calls that were in the wrong class.