feat: let lua scripts detect hup/ctrl-c
This provides a dlua wrapper to crawl_state.seen_hups (which for scripts
now reports both SIGHUP and SIGINT) and uses it in the seed explorer to
react appropriately. Possibly it would be fine to expose this in clua?

More generally, it would be nice if there were a more automated way to
interrupt lua scripts on a signal, but I'm not sure what that is in the
crawl setup, at least to do in a safe/clean way. Right now a stuck rc
script will eventually trigger a SIGABRT in online circumstances if a
regular kill fails, which does interrupt (and not particularly cleanly).
Locally a stuck script usually requires a manual kill.