fix: maybe make some menu code more robust
This code can apparently crash, which isn't surprising given how insane
the type handling is. However, I can't immediately see what is crashing.
Probably this could use a bigger rewrite, but this commit tries to amp
up the explicit type checking a bit so that if it crashes, it'll at
least be more apparent why.