Adjust Xom mutation handling
Remove mutflag::xom, which marked a subset of bad mutations that Xom was
particularly likely to apply. Instead of using a special-cased way of
choosing mutations for Xom's bad mutation action, just use the standard
random mutation codepath but with a slightly higher chance of picking a bad
mutation (a 50/50 split of good/bad, instead of 60/40 for normal random
mutations). This should work out as giving a similar ratio of mutations but
without the special-cased handling of some mutations being extra common.

Closes #2614.