docs: update `menu_colour` documentation
This was out of date in many ways, referring to various removed items,
food, etc., and missing some tags that are actually useful. I've also
tried to clarify the scope of this option as per discussion in #2873.
(As a side note, it's non-ideal that notes are colored with this option,
that should probably get broken out. It also seems like `any` prefixes
shouldn't apply across menu types, or alternatively there should be a
better prefix for indicating "items only".)

I did remove a bunch of the examples from the docs, because they were
outdated and somewhat confusingly redundant. It might be good to have
more examples in here, but they'd need to be a bit more current and less
redundant with the defaults. Maybe someone can pull a few from a
community rc?

Also fixed a few things in the actual defaults:
* runes were missing a color because they are no longer misc items
* removed old annotations related to food, this may cause some
  back-compat errors but that's probably ok at this point.