docs: update `menu_colour` documentation a bit more
`(` and `)` were the keys for displaying wielded and secondary weapons
back in 0.4. This behaviour was moved to `}` in 0.5 (42d8d07399), then
it was assigned to `Ctrl-v` in 0.10 (875129a2f9), and finally it was
removed in 0.11 (ff9fb87a6f). The corresponding part in the options
guide hasn't been updated since 2009.

Also, rip rods.