FIX: Make the "browse character notes" screen obey menu_colour.
The ability to highlight notes in different colours in the "browse
character notes" display (constrolled with the menu_colour option) was
lost with a02aae867c.

This returns it, but with one small change. If a note is too long to
display on a single line of the display, the colour is chosen based on
the whole note, instead of assessing each line of output separately.