fix: further repair work on `menu_colour`
This has always been a weird option, because it is mostly about items,
but has also (at some points) worked for things that aren't items, and
really share no resemblance to items. This is pretty bug-prone (e.g. see
8abc85282e38), and has also just led to some of these options silently
failing over time as people failed to maintain them (see #3056).

This commit refactors the option so that the "any" or no-tag rules do
not apply to any many, but rather any item menu. "any" is
correspondingly renamed to "item" in the docs (though "any" is still
usable for back compatibility). I've also fixed pickup and "resists"
behavior, improved the documentation, and removed a few of these tags.

In particular, spell/ability are broken, have been for a very long time
(from just reading the code, I doubt they worked so long ago as 0.3,
though I didn't actually test), so I've just removed them from the docs.
I'm not opposed to them returning but in modern crawl it'd be a bit
tricky, since they would need to not override the hard-coded rules for
these menus. "notes" also doesn't work but #3056 will fix that (though
to be honest, if this PR didn't exist I'd also just remove that tag).
The help menu stuff probably semi-works (I didn't test carefully),
though I'm really unsure of the value. I don't know if it would be crazy
to make this option item-only, move the `note` one to a different
option, and remove `help`.