fix: color equipped items in use item menus
This was working as intended, in that the default rule explicitly had an
inventory tag (so I guess there was a bug before), and therefore
shouldn't have had a highlight in use item menus. I don't see any
reason to restrict this to just inventory.

Also, document the tag for these menus, and tweak a few more things in
the menu_colour docs.