Make Plasma Beam less extreme
Reduce damage to 25 (from 20) and reduce damage to 2/3rds of its
previous value. It was doing >50% more damage than starburst or iron
shot on a double hit, and those are higher level spells - more than
double fireball's damage. It's louder and less accurate than those
spells, and has worse targeting, but there's a limit to how much
that buys you. Let's smooth out the lows and the highs of the spell
a bit.