Fix Dith webtiles UI
<darkgrey>You can transform into a swirling mass of shadows.(9 HP, Piety---, Max HP drain)

The above description was too long after "Piety" turned into "Piety---",
meaning that our regexp for god powers broke, meaning that we got a
'buggy UI of type describe-god' error.

'Fix' this by shortening the way we describe HP drain in ability costs,
matching how we refer to Drain elsewhere. This is a bit fragile :)