fix: prevent a (harmless?) exception in webtiles
Recent changes make it more typical to send the lobby in an async
fashion, but that means (I think) that the `send_lobby_html` can
occasionally get sequenced after a close call, if a connection opens and
closes immediately for example. This is resulting in a Tornado exception
that is not very informative on the `render_string` call. This changes
adds an early return for the call that would crash.