feat: add a restful endpoint for server version
I think this should be fine to expose publicly (it was previously
visible in the admin panel). This information is tremendously useful to
the subset of the dev team that deals with server issues, but previously
it's been hard to get.

Note that the crawl version string printed here is *not* a game version
per se: it's the version string for what the webtiles server is running.
It would probably be nice to eventually provide a similar endpoint to
get available game versions (which should be doable via the load_games
data) but that's not what this does.