feat: use aiofiles for webtile status file writing if available
This is somewhat experimental. Basically this package does threaded file
IO so as not to block the event loop, which is our current biggest
source of blocking on (non-CAO) servers; currently used only for one
specific case (status file writing). If this works ok I would probably
expand it to be used for ttyrec writing (more technically involved than
this) and milestone tailing.