fix: throttle some webtiles error spam
When the server gets into a bad state and the socket buffers start to
get filled up, on recovery it often emits a very large number of logged
warnings in `on_message` (usually just many "Socket buffer full"
messages). This itself can be pretty IO intensive.  This commit
suppresses all but the first one in a sequence of errors.

Also, hide a common harmless warning that can happen on a disconnect.