Rework the sceptre of Asmodeus
As with the sceptre of Torment, make its base weapon type an eveningstar
so that it's a bit more usable as a melee weapon. Instead of a brand, give
it an extra-high enchantment, and reduce Asmodeus's base damage to
compensate for the better weapon. In addition, turn its activated ability
into a passive one of summoning fiery demons on hit.

It would be nice to come up with a more substantial rework as it probably
still remains the least exciting of the Hell/Pan lord unrands, but making
it usable in melee is hopefully a start!